iot-gateway/requirements.txt
adafruit-io
python-dotenv
pyserial